+++ src/main/app/character/CharacterViewPage.tsx
@@ -48,8 +48,7 @@ const CharacterViewPage: FC<{ character: Character }> = ({ character }) => {
                 </Stack>
                 <Typography
                   level="body-xs"
-                  color={origin === 'praetorian' ? 'warning' : 'primary'}
-                  sx={{ filter: 'brightness(0.6)' }}>
+                  color={origin === 'praetorian' ? 'warning' : 'primary'}>
                   <em>{origin === 'praetorian' ? 'Praetorian' : 'Primal'}</em>
 
                   <Typography sx={{ pl: 0.4 }}>{sex === 'F' ? <Icons.Female/> : <Icons.Male/>}</Typography>